+++ assets/web/src/js/ui/components/Host.jsx
@@ -1,0 +1,9 @@
+export function Host() {
+  return (
+    <div className="me-4 align-self-center">
+      <span className="navbar-text font-monospace text-white">
+        <i className="fa-solid fa-computer"></i> {varnishmon.storage.hostname}
+      </span>
+    </div>
+  );
+}